[Feat] VideoComponets에서 videoMain으로 상태 이동
