+++ src/App.tsx
@@ -1,6 +1,6 @@
 import "./tailwind.css";
 import React from "react";
-import GetData from "@pages/GetData";
+// import GetData from "@pages/GetData";
 import NotFound from "@pages/NotFound";
 import VideoMain from "@pages/VideoMain";
 import VideoDetail from "@pages/VideoDetail";
@@ -13,9 +13,9 @@ function App() {
     <Router>
       <Routes>
         <Route path="/" element={<RootLayout />}>
-          <Route index element={<GetData />} />
+          {/* <Route index element={<GetData />} /> */}
           <Route path="videoDetail" element={<VideoDetail />} />
-          <Route path="videoMain" element={<VideoMain />} />
+          <Route index element={<VideoMain />} />
           <Route path="detail/:id" element={<GetDetailData />} />
           <Route path="*" element={<NotFound />} />
         </Route>
