+++ src/utils/getData.ts
@@ -1,5 +1,4 @@
 import axios from "axios";
-import React, { useEffect, useState } from "react";
 
 async function getData() {
   try {
